backend/requirements.txt
dj-database-url==0.5.0
Django==2.0.7
django-cors-headers==2.4.0
djangorestframework==3.9.4
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==3.3.1
